+++ x/auth/tx/service_test.go
@@ -198,21 +198,22 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPC() {
 		req       *tx.GetTxsEventRequest
 		expErr    bool
 		expErrMsg string
+		expLen    int
 	}{
 		{
 			"nil request",
 			nil,
-			true, "request cannot be nil",
+			true, "request cannot be nil", 0,
 		},
 		{
 			"empty request",
 			&tx.GetTxsEventRequest{},
-			true, "must declare at least one event to search",
+			true, "must declare at least one event to search", 0,
 		},
 		{
 			"request with dummy event",
 			&tx.GetTxsEventRequest{Events: []string{"foobar"}},
-			true, "event foobar should be of the format: {eventType}.{eventAttribute}={value}",
+			true, "event foobar should be of the format: {eventType}.{eventAttribute}={value}", 0,
 		},
 		{
 			"request with order-by",
@@ -220,33 +221,30 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPC() {
 				Events:  []string{bankMsgSendEventAction},
 				OrderBy: tx.OrderBy_ORDER_BY_ASC,
 			},
-			false, "",
+			false, "", 3,
 		},
 		{
 			"without pagination",
 			&tx.GetTxsEventRequest{
 				Events: []string{bankMsgSendEventAction},
 			},
-			false, "",
+			false, "", 3,
 		},
 		{
 			"with pagination",
 			&tx.GetTxsEventRequest{
 				Events: []string{bankMsgSendEventAction},
-				Pagination: &query.PageRequest{
-					CountTotal: false,
-					Offset:     0,
-					Limit:      1,
-				},
+				Page: 2,
+				Limit: 2,
 			},
-			false, "",
+			false, "", 1,
 		},
 		{
 			"with multi events",
 			&tx.GetTxsEventRequest{
 				Events: []string{bankMsgSendEventAction, "message.module='bank'"},
 			},
-			false, "",
+			false, "", 3,
 		},
 	}
 	for _, tc := range testCases {
@@ -260,7 +258,7 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPC() {
 				s.Require().NoError(err)
 				s.Require().GreaterOrEqual(len(grpcRes.Txs), 1)
 				s.Require().Equal("foobar", grpcRes.Txs[0].Body.Memo)
-
+				s.Require().Equal(len(grpcRes.Txs), tc.expLen)
 				// Make sure fields are populated.
 				// ref: https://github.com/cosmos/cosmos-sdk/issues/8680
 				// ref: https://github.com/cosmos/cosmos-sdk/issues/8681
@@ -278,54 +276,55 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPCGateway() {
 		url       string
 		expErr    bool
 		expErrMsg string
+		expLen    int
 	}{
 		{
 			"empty params",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs", val.APIAddress),
 			true,
-			"must declare at least one event to search",
+			"must declare at least one event to search", 0,
 		},
 		{
 			"without pagination",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s", val.APIAddress, bankMsgSendEventAction),
 			false,
-			"",
+			"", 3,
 		},
 		{
 			"with pagination",
-			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&pagination.offset=%d&pagination.limit=%d", val.APIAddress, bankMsgSendEventAction, 0, 10),
+			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&page=%d&limit=%d", val.APIAddress, bankMsgSendEventAction, 2, 2),
 			false,
-			"",
+			"", 1,
 		},
 		{
 			"valid request: order by asc",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=ORDER_BY_ASC", val.APIAddress, bankMsgSendEventAction, "message.module='bank'"),
 			false,
-			"",
+			"", 3,
 		},
 		{
 			"valid request: order by desc",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=ORDER_BY_DESC", val.APIAddress, bankMsgSendEventAction, "message.module='bank'"),
 			false,
-			"",
+			"", 3,
 		},
 		{
 			"invalid request: invalid order by",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s&order_by=invalid_order", val.APIAddress, bankMsgSendEventAction, "message.module='bank'"),
 			true,
-			"is not a valid tx.OrderBy",
+			"is not a valid tx.OrderBy", 0,
 		},
 		{
 			"expect pass with multiple-events",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s&events=%s", val.APIAddress, bankMsgSendEventAction, "message.module='bank'"),
 			false,
-			"",
+			"", 3,
 		},
 		{
 			"expect pass with escape event",
 			fmt.Sprintf("%s/cosmos/tx/v1beta1/txs?events=%s", val.APIAddress, "message.action%3D'/cosmos.bank.v1beta1.MsgSend'"),
 			false,
-			"",
+			"", 3,
 		},
 	}
 	for _, tc := range testCases {
@@ -341,6 +340,7 @@ func (s IntegrationTestSuite) TestGetTxEvents_GRPCGateway() {
 				s.Require().GreaterOrEqual(len(result.Txs), 1)
 				s.Require().Equal("foobar", result.Txs[0].Body.Memo)
 				s.Require().NotZero(result.TxResponses[0].Height)
+				s.Require().Equal(len(result.Txs), tc.expLen)
 			}
 		})
 	}